Use explicit type declarations for uint256

+++ contracts/EIP777/UnstructuredTokenBase.sol
@@ -12,7 +12,7 @@ import "../ownership/UnstructuredOwnable.sol";
 
 /**
 * @title UnstructuredTokenBase
-* @dev 
+* @dev
 *   This is an ERC777/20 compatible token base with upgradable properties
 *   that use a proxy/dispatch (DELEGATECALL) variation that allows for
 *   storage changes over time. IE, you can define new vars in new versions
@@ -20,19 +20,19 @@ import "../ownership/UnstructuredOwnable.sol";
 contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Implementer {
 
   using SafeMath for uint256;
-  
+
   string private mName;
   string private mSymbol;
   uint256 private mGranularity;
   uint256 private mTotalSupply;
   bool internal _initialized;
-  
+
   bool private mErc20compatible;
-  
+
   mapping(address => uint) private mBalances;
   mapping(address => mapping(address => bool)) private mAuthorized;
   mapping(address => mapping(address => uint256)) private mAllowed;
-  
+
   ///  @notice This modifier is applied to erc20 obsolete methods that are
   ///  implemented only to maintain backwards compatibility. When the erc20
   ///  compatibility is disabled, this methods will fail.
@@ -40,10 +40,10 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     require(mErc20compatible);
     _;
   }
-  
-  
+
+
   constructor () public { }
-  
+
   function initialize(
     string _name,
     string _symbol,
@@ -59,9 +59,9 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     mErc20compatible = true;
     require(_granularity >= 1);
     mGranularity = _granularity;
-  
+
     setOwner(_owner);
-    
+
     setIntrospectionRegistry(_eip820RegistryAddr);
     setInterfaceImplementation("IEIP777", this);
     setInterfaceImplementation("Ierc20", this);
@@ -75,61 +75,61 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   function initialized() public view returns(bool) {
     return _initialized;
   }
-  
+
   /// @return the name of the token
   function name() public view returns (string) { return mName; }
-  
+
   /// @return the symbol of the token
   function symbol() public view returns(string) { return mSymbol; }
-  
+
   /// @return the granularity of the token
   function granularity() public view returns(uint256) { return mGranularity; }
-  
+
   /// @return the total supply of the token
   function totalSupply() public view returns(uint256) { return mTotalSupply; }
-  
+
   /// @notice For Backwards compatibility
   /// @return The decimls of the token. Forced to 18 in ERC777.
   function decimals() public erc20 view returns (uint8) { return uint8(18); }
-  
+
   /// @notice Return the account balance of some account
   /// @param _tokenHolder Address for which the balance is returned
   /// @return the balance of `_tokenAddress`.
   function balanceOf(address _tokenHolder) public view returns (uint256) {
     return mBalances[_tokenHolder];
   }
-  
+
   /// @notice Disables the ERC-20 interface. This function can only be called
   ///  by the owner.
   function disableERC20() public onlyOwner {
     mErc20compatible = false;
     setInterfaceImplementation("Ierc20", 0x0);
   }
-  
+
   /// @notice Re enables the ERC-20 interface. This function can only be called
   ///  by the owner.
   function enableERC20() public onlyOwner {
     mErc20compatible = true;
     setInterfaceImplementation("Ierc20", this);
   }
-  
+
   /// @notice ERC20 backwards compatible transfer.
   /// @param _to The address of the recipient
   /// @param _value The amount of tokens to be transferred
   /// @return `true`, if the transfer can't be done, it should fail.
   function transfer(address _to, uint256 _value) public erc20 returns (bool success) {
     doSend(
-      msg.sender, 
-      _to, 
-      _value, 
-      "", 
-      msg.sender, 
-      "", 
+      msg.sender,
+      _to,
+      _value,
+      "",
+      msg.sender,
+      "",
       false
     );
     return true;
   }
-  
+
   /// @notice ERC20 backwards compatible transferFrom.
   /// @param _from The address holding the tokens being transferred
   /// @param _to The address of the recipient
@@ -137,21 +137,21 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   /// @return `true`, if the transfer can't be done, it should fail.
   function transferFrom(address _from, address _to, uint256 _value) public erc20 returns (bool success) {
     require(_value <= mAllowed[_from][msg.sender]);
-    
+
     // Cannot be after doSend because of tokensReceived re-entry
     mAllowed[_from][msg.sender] = mAllowed[_from][msg.sender].sub(_value);
     doSend(
-      _from, 
-      _to, 
-      _value, 
-      "", 
-      msg.sender, 
-      "", 
+      _from,
+      _to,
+      _value,
+      "",
+      msg.sender,
+      "",
       false
     );
     return true;
   }
-  
+
   /// @notice ERC20 backwards compatible approve.
   ///  `msg.sender` approves `_spender` to spend `_value` tokens on its behalf.
   /// @param _spender The address of the account able to transfer the tokens
@@ -162,7 +162,7 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     emit Approval(msg.sender, _spender, _value);
     return true;
   }
-  
+
   /// @notice ERC20 backwards compatible allowance.
   ///  This function makes it easy to read the `allowed[]` map
   /// @param _owner The address of the account that owns the token
@@ -172,38 +172,38 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   function allowance(address _owner, address _spender) public erc20 view returns (uint256 remaining) {
     return mAllowed[_owner][_spender];
   }
-  
+
   /// @notice Send `_value` amount of tokens to address `_to`
   /// @param _to The address of the recipient
   /// @param _value The amount of tokens to be sent
   function send(address _to, uint256 _value) public {
     doSend(
-      msg.sender, 
-      _to, 
-      _value, 
-      "", 
-      msg.sender, 
-      "", 
+      msg.sender,
+      _to,
+      _value,
+      "",
+      msg.sender,
+      "",
       true
     );
   }
-  
+
   //todo jaycen are both send funcs needed for some backwards compat reason? check latest 777 draft
   /// @notice Send `_value` amount of tokens to address `_to` passing `_userData` to the recipient
   /// @param _to The address of the recipient
   /// @param _value The amount of tokens to be sent
   function send(address _to, uint256 _value, bytes _userData) public {
     doSend(
-      msg.sender, 
-      _to, 
-      _value, 
-      _userData, 
-      msg.sender, 
-      "", 
+      msg.sender,
+      _to,
+      _value,
+      _userData,
+      msg.sender,
+      "",
       true
     );
   }
-  
+
   /// @notice Authorize a third party `_operator` to manage (send) `msg.sender`'s tokens.
   /// @param _operator The operator that wants to be Authorized
   //todo jaycen pretty sure overloads dont work so calling the following is impossible
@@ -221,17 +221,17 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     //todo jaycen is the following needed, It is not in the spec
     mAuthorized[_operator][msg.sender] = true;
     callOperator(
-      _operator, 
-      msg.sender, 
-      _operator, 
-      _value, 
-      "0x0", 
-      "0x0", 
+      _operator,
+      msg.sender,
+      _operator,
+      _value,
+      "0x0",
+      "0x0",
       false
     );
     emit AuthorizedOperator(_operator, msg.sender);
   }
-  
+
   /// @notice Revoke a third party `_operator`'s rights to manage (send) `msg.sender`'s tokens.
   /// @param _operator The operator that wants to be Revoked
   function revokeOperator(address _operator) public {
@@ -242,7 +242,7 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     mAuthorized[_operator][msg.sender] = false;
     emit RevokedOperator(_operator, msg.sender);
   }
-  
+
   /// @notice Check whether the `_operator` address is allowed to manage the tokens held by `_tokenHolder` address.
   /// @param _operator address to check if it has the right to manage the tokens
   /// @param _tokenHolder address which holds the tokens to be managed
@@ -250,7 +250,7 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   function isOperatorFor(address _operator, address _tokenHolder) public view returns (bool) {
     return _operator == _tokenHolder || mAuthorized[_operator][_tokenHolder];
   }
-  
+
   /// @notice Send `_value` amount of tokens on behalf of the address `from` to the address `to`.
   /// @param _from The address holding the tokens being sent
   /// @param _to The address of the recipient
@@ -258,11 +258,11 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   /// @param _userData Data generated by the user to be sent to the recipient
   /// @param _operatorData Data generated by the operator to be sent to the recipient
   function operatorSend(
-    address _operator, 
-    address _from, 
-    address _to, 
-    uint256 _value, 
-    bytes _userData, 
+    address _operator,
+    address _from,
+    address _to,
+    uint256 _value,
+    bytes _userData,
     bytes _operatorData
   ) public {
     //todo jaycen for somereason the following two lines werent included in the 777
@@ -272,15 +272,15 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     require(isOperatorFor(_operator, _from));
     doSend(
       _from,
-      _to, 
-      _value, 
-      _userData, 
-      msg.sender, 
-      _operatorData, 
+      _to,
+      _value,
+      _userData,
+      msg.sender,
+      _operatorData,
       false
     );
   }
-  
+
   /// @notice Send `_value` amount of tokens on behalf of the address `from` to the address `to`.
   /// @param _from The address holding the tokens being sent
   /// @param _to The address of the recipient
@@ -288,23 +288,23 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   /// @param _userData Data generated by the user to be sent to the recipient
   /// @param _operatorData Data generated by the operator to be sent to the recipient
   function operatorSend(
-    address _from, 
-    address _to, 
-    uint256 _value, 
-    bytes _userData, 
+    address _from,
+    address _to,
+    uint256 _value,
+    bytes _userData,
     bytes _operatorData
   ) public {
     // for backwards compatibility (only should be called when callOperator NOT being used)
     operatorSend(
-      msg.sender, 
-      _from, 
-      _to, 
-      _value, 
-      _userData, 
+      msg.sender,
+      _from,
+      _to,
+      _value,
+      _userData,
       _operatorData
     );
   }
-  
+
   /// @notice Burns `_value` tokens from `_tokenHolder`
   ///  Sample burn function to showcase the use of the `Burnt` event.
   /// @param _tokenHolder The address that will lose the tokens
@@ -312,16 +312,16 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
   function burn(address _tokenHolder, uint256 _value) public onlyOwner {
     requireMultiple(_value);
     require(balanceOf(_tokenHolder) >= _value);
-    
+
     mBalances[_tokenHolder] = mBalances[_tokenHolder].sub(_value);
     mTotalSupply = mTotalSupply.sub(_value);
-    
+
     emit Burnt(_tokenHolder, _value);
     if (mErc20compatible) {
       emit Transfer(_tokenHolder, 0x0, _value);
     }
   }
-  
+
   /// @notice Generates `_value` tokens to be assigned to `_tokenHolder`
   /// @param _tokenHolder The address that will be assigned the new tokens
   /// @param _value The quantity of tokens generated
@@ -331,34 +331,34 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     requireMultiple(_value);
     mTotalSupply = mTotalSupply.add(_value);
     mBalances[_tokenHolder] = mBalances[_tokenHolder].add(_value);
-    
+
     callRecipent(
-      msg.sender, 
-      0x0, 
-      _tokenHolder, 
-      _value, 
-      "", 
-      _operatorData, 
+      msg.sender,
+      0x0,
+      _tokenHolder,
+      _value,
+      "",
+      _operatorData,
       true
     );
-    
+
     emit Minted(
-      _tokenHolder, 
-      _value, 
-      msg.sender, 
+      _tokenHolder,
+      _value,
+      msg.sender,
       _operatorData
     );
     if (mErc20compatible) {
       emit Transfer(0x0, _tokenHolder, _value);
     }
   }
-  
+
   /// @notice Internal function that ensures `_value` is multiple of the granularity
   /// @param _value The quantity that want's to be checked
   function requireMultiple(uint256 _value) internal view {
     require(_value > 0 && _value.div(mGranularity).mul(mGranularity) == _value);
   }
-  
+
   /// @notice Check whether an address is a regular address or not.
   /// @param _addr Address of the contract that has to be checked
   /// @return `true` if `_addr` is a regular address (not a contract)
@@ -366,11 +366,11 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     if (_addr == 0) {
       return false;
     }
-    uint size;
+    uint256 size;
     assembly { size := extcodesize(_addr) } // solium-disable-line security/no-inline-assembly
     return size == 0;
   }
-  
+
   /// @notice Helper function actually performing the sending of tokens.
   /// @param _from The address holding the tokens being sent
   /// @param _to The address of the recipient
@@ -393,23 +393,23 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     requireMultiple(_value);
     require(_to != address(0));          // forbid sending to 0x0 (=burning)
     require(mBalances[_from] >= _value); // ensure enough funds
-    
+
     mBalances[_from] = mBalances[_from].sub(_value);
     mBalances[_to] = mBalances[_to].add(_value);
-    
+
     callRecipent(
-      _operator, 
-      _from, 
-      _to, 
-      _value, 
-      _userData, 
-      _operatorData, 
+      _operator,
+      _from,
+      _to,
+      _value,
+      _userData,
+      _operatorData,
       _preventLocking
     );
-    
+
     emit Sent(
-      _from, 
-      _to, 
+      _from,
+      _to,
       _value,
       _userData,
       _operator,
@@ -419,7 +419,7 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
       emit Transfer(_from, _to, _value);
     }
   }
-  
+
   /// @notice Helper function that checks for IEIP777TokensRecipient on the recipient and calls it.
   ///  May throw according to `_preventLocking`
   /// @param _operator The address authorized to send tokens
@@ -444,18 +444,18 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     address recipientImplementation = interfaceAddr(_to, "IEIP777TokensRecipient");
     if (recipientImplementation != 0) {
       IEIP777TokensRecipient(recipientImplementation).tokensReceived(
-        _operator, 
-        _from, 
-        _to, 
-        _value, 
-        _userData, 
+        _operator,
+        _from,
+        _to,
+        _value,
+        _userData,
         _operatorData
       );
     } else if (_preventLocking) {
       require(isRegularAddress(_to));
     }
   }
-  
+
   /// @notice Helper function that checks for IEIP777TokensOperator on the recipient and calls it.
   ///  May throw according to `_preventLocking`
   /// @param _from The address holding the tokens being sent
@@ -480,10 +480,10 @@ contract UnstructuredTokenBase is UnstructuredOwnable, Ierc20, IEIP777, EIP820Im
     if (recipientImplementation != 0) {
       IEIP777TokensOperator(recipientImplementation).madeOperatorForTokens(
         _operator,
-        _from, 
-        _to, 
-        _value, 
-        _userData, 
+        _from,
+        _to,
+        _value,
+        _userData,
         _operatorData
       );
     } else if (_preventLocking) {
