+++ contracts/commodity/ICommoditySender.sol
@@ -5,7 +5,7 @@ interface ICommoditySender {
     address operator,
     address from,
     address to,
-    uint id,
+    uint256 id,
     bytes userData,
     bytes operatorData
   ) public;
