+++ contracts/commodity/SplittableCommodity.sol
@@ -10,7 +10,7 @@ contract SplittableCommodity is MintableCommodity {
 
   //todo jaycen implement a version where owner is splitting vs market is splitting
   //todo jaycen, maybe if we dont store data for misc in the new commodity and instead rely on front end to retrieve based on parentid to save gas
-  function split(uint _tokenId, address _to, uint256 _amount) public {
+  function split(uint256 _tokenId, address _to, uint256 _amount) public {
     commodities[_tokenId].value =  commodities[_tokenId].value.sub(_amount);
     uint64 newCommodityIndex = mint(
       _to,
